refactor [James]: refactor events route
